implement uderID in rules firebase

+++ lib/repositories/items_repository.dart
@@ -1,9 +1,17 @@
 import 'package:cantwait28/models/item_model.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
 class ItemsRepository {
   Stream<List<ItemModel>> getItemsStream() {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik jest nie zalogowany');
+    }
+
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
         .orderBy('release_date')
         .snapshots()
@@ -20,7 +28,13 @@ class ItemsRepository {
   }
 
   Future<ItemModel> get({required String documentID}) async {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik jest nie zalogowany');
+    }
     final doc = await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
         .doc(documentID)
         .get();
@@ -33,8 +47,14 @@ class ItemsRepository {
   }
 
   Future<void> remove({required String documentID}) {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik jest nie zalogowany');
+    }
     {
       return FirebaseFirestore.instance
+          .collection('users')
+          .doc(userID)
           .collection('items')
           .doc(documentID)
           .delete();
@@ -46,7 +66,15 @@ class ItemsRepository {
     String imageURL,
     DateTime releaseDate,
   ) async {
-    await FirebaseFirestore.instance.collection('items').add(
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik jest nie zalogowany');
+    }
+    await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
+        .collection('items')
+        .add(
       {
         'title': title,
         'image_url': imageURL,
